Opacus-PRV/pyproject.toml
[tool.poetry]
name = "opacus"
version = "0.2.0"
description = ""
authors = ["Roy Rinberg <royrinberg@gmail.com>"]
readme = "README.md"

packages = [ {include = "opacus"}, ]


[tool.poetry.dependencies]
python = ">=3.8, <3.10"
numpy = "^1.22.0"
torch = "^1.12.0"
scipy = "^1.8.1"
opt-einsum = "^3.3.0"

ipykernel = "^6.15.1"
torchvision = "^0.13.0"
tqdm = "^4.64.0"
pylint = "^2.14.4"

PyYAML = "^6.0"
toml = "^0.10.2"
kymatio = "^0.3.0"
transformers = "^4.31.0"
tensorflow = "^2.13.0"
keras = "^2.13.1"


[tool.poetry.group.dev.dependencies]

jupyter = "^1.0.0"
yapf = "^0.32.0"
black = "^22.6.0"
bpython = "^0.22.1"
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
